data digitizer from printed papers
This repo is not a final pack but can be used with some minor revisions based on your use case.
It reads a paper print to fill out the fields necessary for AI training in Ophtalmology. (This is not an annotation tool. It is used for data mining of hardcopies.)